Rename ground2 symbol to rail
