+++ symbols/rail_up.ts
@@ -1,16 +1,15 @@
-import svgJson from "assets/generated/ground2.json"
+import svgJson from "assets/generated/rail.json"
 import { modifySymbol } from "drawing/modify-symbol/modify-symbol"
 
-const { paths, texts, bounds, refblocks, circles } = svgJson
+const { paths, texts, bounds, refblocks } = svgJson
 export default modifySymbol({
   primitives: [
     ...Object.values(paths),
-    ...Object.values(circles),
     {
       type: "text",
       text: "{REF}",
       x: -0.14,
-      y: -0.0,
+      y: 0,
     },
   ] as any,
   ports: [
@@ -18,7 +17,7 @@ export default modifySymbol({
   ],
   center: { x: bounds.centerX, y: bounds.centerY },
 })
-  .rotateRightFacingSymbol("up")
+  .rotateRightFacingSymbol("down")
   .labelPort("left1", ["1"])
-  .changeTextAnchor("{REF}", "middle_top")
+  .changeTextAnchor("{REF}", "middle_bottom")
   .build()
